add sync service + hierarchy
