+++ packages/syft/src/syft/node/node.py
@@ -88,6 +88,7 @@ from ..service.response import SyftError
 from ..service.service import AbstractService
 from ..service.service import ServiceConfigRegistry
 from ..service.service import UserServiceConfigRegistry
+from ..service.sync.sync_service import SyncService
 from ..service.settings.settings import NodeSettingsV2
 from ..service.settings.settings_service import SettingsService
 from ..service.settings.settings_stash import SettingsStash
@@ -357,6 +358,7 @@ class Node(AbstractNode):
                 SyftWorkerImageService,
                 SyftWorkerPoolService,
                 SyftImageRegistryService,
+                SyncService,
             ]
             if services is None
             else services
@@ -934,6 +936,7 @@ class Node(AbstractNode):
                 SyftWorkerImageService,
                 SyftWorkerPoolService,
                 SyftImageRegistryService,
+                SyncService,
             ]
 
             if OBLV:
